Fix for split log lines
Sometimes a log line gets split after item name, the rest of the line goes to the next line. Here is a fix. Cause unknown at this stage if it hasnt been happening leave FIX_BROKEN_LOG_LINES = false;

If it does happen (split line) that line wont be inserted and it will throw and error.