New item id implementation
Why?: Old item id system was assigned sequentially from the trader config file. The first item listed was 1, second listed was 2 etc. If the items listed got changed around then the item ids could then be assigned to another classname thus wrecking the data.

This new system checks the item table for the classname, if it isnt found it will create an entry and returned the incremented id if the classname is found it will return the id which was detirmined the first time this item was inserted.